requirements.txt
google-generativeai>=0.8.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0
networkx>=3.2
aiohttp>=3.9.0
aiofiles>=23.2.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
tiktoken>=0.5.0
PyYAML>=6.0.1
tenacity>=8.2.0
jsonschema>=4.20.0
